unify nat extration test into single test
